Rename package email to sender
