+++ internal/sender/message.go
@@ -1,4 +1,4 @@
-package email
+package sender
 
 import (
 	"bytes"
