+++ internal/sender/message_test.go
@@ -1,4 +1,4 @@
-package email
+package sender
 
 import (
 	"testing"
