+++ internal/sender/stub.go
@@ -1,4 +1,4 @@
-package email
+package sender
 
 import (
 	"crypto/tls"
